backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.2.0

# File handling dependencies
aiofiles==23.2.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2